Rename WalletsActivty to AssetsActivity
