+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme" />
         <activity
-            android:name=".activities.WalletsActivity"
+            android:name=".activities.AssetsActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme" />
         <activity
