+++ app/src/main/res/layout/activity_base_popup.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.WalletsActivity">
+    tools:context=".activities.AssetsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
