+++ app/src/main/res/layout/content_assets_activity.xml
